Refactor code, include custom checkbox, include steps in markdown
