+++ packages/gh-page/src/markdown/Markdown.tsx
@@ -1,17 +1,20 @@
-import ReactMarkdown from "react-markdown";
-import { ReactMarkdownOptions } from "react-markdown/lib/react-markdown";
-import rehypeHighlight from "rehype-highlight";
+import ReactMarkdown from 'react-markdown'
+import { ReactMarkdownOptions } from 'react-markdown/lib/react-markdown'
+import rehypeHighlight from 'rehype-highlight'
 
-import { colorHarmoniesMarkdown } from "./color-harmonies-markdown";
-import { cssStyleDeclarationMarkdown } from "./css-style-declaration-markdown";
-import { darkenLightenMarkdown } from "./darken-lighten-markdown";
-import { overviewMarkdown } from "./overview-markdown";
+import { colorHarmoniesMarkdown } from './color-harmonies-markdown'
+import { cssStyleDeclarationMarkdown } from './css-style-declaration-markdown'
+import { darkenLightenMarkdown } from './darken-lighten-markdown'
+import { tldrMarkdown } from './tlrl-markdown'
+import { overviewMarkdown } from './why-markdown'
 
 export const Markdown = ({ className, children }: ReactMarkdownOptions) => (
   <ReactMarkdown rehypePlugins={[rehypeHighlight]} className={className}>
     {children}
   </ReactMarkdown>
 );
+
+export const TlDrMarkdown = () => <Markdown>{tldrMarkdown}</Markdown>;
 
 export const OverviewMarkdown = () => <Markdown>{overviewMarkdown}</Markdown>;
