+++ packages/gh-page/src/markdown/css-style-declaration-markdown.tsx
@@ -1,8 +1,3 @@
-import hexToHsl from 'hex-to-hsl'
-import { useEffect, useState } from 'react'
-
-import { setTheme } from '@tame-your-theme/css-style-declaration'
-
 export const cssStyleDeclarationMarkdown = `## Functions in \`@tame-your-theme/css-style-declaration\`
 
 The idea of this package is to benefit from the native api from [CSS Object Model](https://developer.mozilla.org/en-US/docs/Web/API/CSS_Object_Model) to change the css variables in runtime.
@@ -34,155 +29,4 @@ export type GetHueSaturationAndLightnessType = (value: string) => {
 \`\`\`
 
 In the example below the lib being used to get these values is [hex-to-hsl](https://www.npmjs.com/package/hex-to-hsl), but one can use any lib or even create the logic to get the values from a hexadecimal color. The following example uses the \`setTheme\` function to change the theme dynamically. To achieve this change, it's necessary to have a proper css set in place in the components.
-
-\`\`\`scss
-.themed {
-  background-color: var(--themed-div-background-color);
-  color: var(--themed-div-background-contrast-color);
-}
-
-.themed h1 {
-  color: var(--themed-div-primary-color);
-}
-\`\`\`
-
-\`\`\`ts
-import hexToHsl from "hex-to-hsl";
-import { setTheme } from "@tame-your-theme/css-style-declaration";
-
-const getHueSaturationAndLightness = (value: string) => {
-  const [hue, saturation, lightness] = hexToHsl(value);
-
-  return {
-    hue,
-    saturation,
-    lightness,
-  };
-};
-
-export const whiteTheme = [
-  {
-    name: "themed-div-background-contrast-color",
-    value: "#212121",
-  },
-  {
-    name: "themed-div-background-color",
-    value: "#ffffff",
-  },
-  {
-    name: "themed-div-primary-color",
-    value: "#ffa500",
-  },
-];
-
-export const darkTheme = [
-  {
-    name: "themed-div-background-contrast-color",
-    value: "#FFFFFF",
-  },
-  {
-    name: "themed-div-background-color",
-    value: "#292d33",
-  },
-  {
-    name: "themed-div-primary-color",
-    value: "#FFB6C1",
-  },
-];
-
-const getHueSaturationAndLightness = (value: string) => {
-  const [hue, saturation, lightness] = hexToHsl(value);
-
-  return {
-    hue,
-    saturation,
-    lightness,
-  };
-};
-
-export const ThemedDiv = () => {
-  const [currentTheme, setCurrentTheme] = useState(whiteTheme);
-
-  useEffect(() => {
-    setTheme({
-      variables: currentTheme,
-      getHueSaturationAndLightness,
-    });
-  }, [currentTheme]);
-
-  const onCheckChange = (event) => {
-    setCurrentTheme(event.target.checked ? darkTheme : whiteTheme);
-  };
-
-  return (
-    <div className="themed">
-      <input onChange={onCheckChange} type="checkbox" /> Dark Theme?
-      <h1>This is a title which will change color</h1>
-    </div>
-  );
-};
-
-\`\`\`
 `;
-
-export const whiteTheme = [
-  {
-    name: "themed-div-background-contrast-color",
-    value: "#212121",
-  },
-  {
-    name: "themed-div-background-color",
-    value: "#ffffff",
-  },
-  {
-    name: "themed-div-primary-color",
-    value: "#ffa500",
-  },
-];
-
-export const darkTheme = [
-  {
-    name: "themed-div-background-contrast-color",
-    value: "#FFFFFF",
-  },
-  {
-    name: "themed-div-background-color",
-    value: "#292d33",
-  },
-  {
-    name: "themed-div-primary-color",
-    value: "#FFB6C1",
-  },
-];
-
-const getHueSaturationAndLightness = (value: string) => {
-  const [hue, saturation, lightness] = hexToHsl(value);
-
-  return {
-    hue,
-    saturation,
-    lightness,
-  };
-};
-
-export const ThemedDiv = () => {
-  const [currentTheme, setCurrentTheme] = useState(whiteTheme);
-
-  useEffect(() => {
-    setTheme({
-      variables: currentTheme,
-      getHueSaturationAndLightness,
-    });
-  }, [currentTheme]);
-
-  const onCheckChange = (event) => {
-    setCurrentTheme(event.target.checked ? darkTheme : whiteTheme);
-  };
-
-  return (
-    <div className="themed">
-      <input onChange={onCheckChange} type="checkbox" /> Dark Theme?
-      <h1>This is a title which will change color</h1>
-    </div>
-  );
-};
