+++ packages/gh-page/src/scss/custom-checkbox.scss
@@ -1,0 +1,57 @@
+.custom-checkbox-container {
+  position: relative;
+  margin-bottom: 12px;
+  cursor: pointer;
+  font-size: 22px;
+  user-select: none;
+
+  input {
+    position: absolute;
+    opacity: 0;
+    cursor: pointer;
+
+    &:checked ~ .checkmark {
+      background-color: var(--themed-div-primary-color);
+
+      &:after {
+        display: block;
+      }
+    }
+  }
+
+  .checkmark {
+    position: absolute;
+    top: 0;
+    left: 205px;
+    height: 25px;
+    width: 25px;
+    background-color: #eee;
+    border: 1px solid
+      change-lightness(--themed-div-background-contrast-color, 30%);
+
+    &:after {
+      content: "";
+      position: absolute;
+      display: none;
+    }
+
+    &:after {
+      left: 9px;
+      top: 5px;
+      width: 5px;
+      height: 10px;
+      border: solid;
+      border-color: var(--themed-div-background-color);
+      border-width: 0 3px 3px 0;
+      transform: rotate(45deg);
+
+      &:hover {
+        border-color: var(--themed-div-background-contrast-color);
+      }
+    }
+  }
+
+  &:hover input ~ .checkmark {
+    background-color: var(--themed-div-primary-color-hover);
+  }
+}
