+++ packages/gh-page/src/scss/guidelines.scss
